孩子们，ACT宙能似乎回来了但没回来

+++ Content/Projectiles/Rogue/Celestus2.cs
@@ -45,8 +45,8 @@ namespace CalamityInheritance.Content.Projectiles.Rogue
             if (Projectile.timeLeft < 85)
             {
                 byte b2 = (byte)(Projectile.timeLeft * 3);
-                byte a2 = (byte)(100f * ((float)b2 / 255f));
-                return new Color((int)b2, (int)b2, (int)b2, (int)a2);
+                byte a2 = (byte)(100f * (b2 / 255f));
+                return new Color(b2, b2, b2, a2);
             }
             return new Color(255, 255, 255, 100);
         }
